refactor: folder structure, file renaming
